Move `KmipServerState` to `cascaded::hsm`
