+++ src/units/http_server.rs
@@ -24,7 +24,6 @@ use domain::utils::base64;
 use domain_kmip::ConnectionSettings;
 use domain_kmip::dep::kmip::client::pool::ConnectionManager;
 use serde::Deserialize;
-use serde::Serialize;
 use tokio::net::TcpListener;
 use tokio::task::JoinSet;
 use tracing::{debug, error, info, warn};
@@ -36,6 +35,7 @@ use crate::api::*;
 use crate::center;
 use crate::center::Center;
 use crate::center::get_zone;
+use crate::hsm::KmipServerState;
 use crate::loader;
 use crate::manager::Terminated;
 use crate::policy::AutoConfig;
@@ -1469,75 +1469,6 @@ impl HttpServer {
 }
 
 //------------ HttpServer Handler for /kmip ----------------------------------
-
-/// Non-sensitive KMIP server settings to be persisted.
-///
-/// Sensitive details such as certificates and credentials should be stored
-/// separately.
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct KmipServerState {
-    pub server_id: String,
-    pub ip_host_or_fqdn: String,
-    pub port: u16,
-    pub insecure: bool,
-    pub connect_timeout: Duration,
-    pub read_timeout: Duration,
-    pub write_timeout: Duration,
-    pub max_response_bytes: u32,
-    pub key_label_prefix: Option<String>,
-    pub key_label_max_bytes: u8,
-    pub has_credentials: bool,
-}
-
-impl From<HsmServerAdd> for KmipServerState {
-    fn from(srv: HsmServerAdd) -> Self {
-        KmipServerState {
-            server_id: srv.server_id,
-            ip_host_or_fqdn: srv.ip_host_or_fqdn,
-            port: srv.port,
-            insecure: srv.insecure,
-            connect_timeout: srv.connect_timeout,
-            read_timeout: srv.read_timeout,
-            write_timeout: srv.write_timeout,
-            max_response_bytes: srv.max_response_bytes,
-            key_label_prefix: srv.key_label_prefix,
-            key_label_max_bytes: srv.key_label_max_bytes,
-            has_credentials: srv.username.is_some(),
-        }
-    }
-}
-
-impl From<KmipServerState> for api::KmipServerState {
-    fn from(value: KmipServerState) -> Self {
-        let KmipServerState {
-            server_id,
-            ip_host_or_fqdn,
-            port,
-            insecure,
-            connect_timeout,
-            read_timeout,
-            write_timeout,
-            max_response_bytes,
-            key_label_prefix,
-            key_label_max_bytes,
-            has_credentials,
-        } = value;
-
-        Self {
-            server_id,
-            ip_host_or_fqdn,
-            port,
-            insecure,
-            connect_timeout,
-            read_timeout,
-            write_timeout,
-            max_response_bytes,
-            key_label_prefix,
-            key_label_max_bytes,
-            has_credentials,
-        }
-    }
-}
 
 impl HttpServer {
     async fn kmip_server_add(
